Remove DEBUG.ASSERT which stopped NCrunch running

+++ libraries/KustoLoco.Core/Evaluation/BuiltIns/Operators/MultiplyFunction.cs
@@ -6,7 +6,7 @@ namespace KustoLoco.Core.Evaluation.BuiltIns.Impl;
 [KustoImplementation(Keyword = "Operators.Multiply")]
 internal partial class MultiplyFunction
 {
-    private static long IntImpl(int a, int b) => a * b;
+    private static long IntImpl(int a, int b) => a * (long)b;
     private static long LongImpl(long a, long b) => a * b;
     private static double DoubleImpl(double a, double b) => a * b;
     private static decimal DecimalImpl(decimal a, decimal b) => a * b;
